adding / fixing sticker / image model

+++ lib/src/api/matrix_client_api/r0/model/event/content/m_room/message/content/m_image/thumbnail_info.dart
@@ -4,10 +4,10 @@ part 'thumbnail_info.g.dart';
 
 @JsonSerializable(nullable: true, )
 class ThumbnailInfo {
-  final int h;
-  final int w;
+  final double h;
+  final double w;
   final String mimetype;
-  final int size;
+  final double size;
 
   ThumbnailInfo({
     this.h,
